# fasofree-backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Créer un compte utilisateur",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Se connecter et obtenir un JWT",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtenir le profil utilisateur connecté",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/businesses": {
      "post": {
        "operationId": "BusinessesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Créer un commerce",
        "tags": [
          "Businesses"
        ]
      },
      "get": {
        "operationId": "BusinessesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lister tous les commerces (super administrateur)",
        "tags": [
          "Businesses"
        ]
      }
    },
    "/api/v1/businesses/nearby": {
      "get": {
        "operationId": "BusinessesController_findNearby",
        "parameters": [
          {
            "name": "latitude",
            "required": true,
            "in": "query",
            "schema": {
              "example": 12.3714,
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": true,
            "in": "query",
            "schema": {
              "example": -1.5197,
              "type": "number"
            }
          },
          {
            "name": "radiusInKm",
            "required": false,
            "in": "query",
            "description": "Rayon de recherche, en kilomètres",
            "schema": {
              "default": 5,
              "example": 5,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Rechercher les commerces à proximité",
        "tags": [
          "Businesses"
        ]
      }
    },
    "/api/v1/businesses/{id}": {
      "get": {
        "operationId": "BusinessesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtenir un commerce et son catalogue",
        "tags": [
          "Businesses"
        ]
      },
      "patch": {
        "operationId": "BusinessesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mettre à jour les paramètres d'un commerce",
        "tags": [
          "Businesses"
        ]
      }
    },
...
    "/api/v1/products/business/{businessId}": {
      "get": {
        "operationId": "ProductsController_findByBusiness",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lister les produits d’un commerce",
        "tags": [
          "Products"
        ]
      }
    },
...
    "/api/v1/analytics/business/{businessId}": {
      "get": {
        "operationId": "AnalyticsController_getBusinessOverview",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir les indicateurs d’un commerce",
        "tags": [
          "Analytics"
        ]
      }
    },
...
    "/api/v1/chat/{orderId}": {
      "get": {
        "operationId": "ChatController_getChatHistory",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historique récupéré"
          },
          "401": {
            "description": "Utilisateur non authentifié"
          },
          "403": {
            "description": "Non participant au canal"
          },
          "404": {
            "description": "Commande introuvable"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Historique du chat éphémère d'une commande (canal merchant|driver)",
        "tags": [
          "Chat"
        ]
      }
    },
...
  },
  "info": {
    "title": "FasoFree API",
    "description": "Spécification OpenAPI officielle du backend FasoFree - Marketplace & Livraison à Ouagadougou",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "Serveur courant"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Saisissez votre token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Aminata Ouédraogo"
          },
          "email": {
            "type": "string",
            "example": "aminata@fasofree.bf",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "example": "+22670000000"
          },
          "password": {
            "type": "string",
            "example": "MotDePasseFort123!",
            "format": "password"
          },
          "referralCode": {
            "type": "string",
            "example": "AMINATA-7F3A"
          }
        },
        "required": [
          "fullName",
          "email",
          "phone",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "client@fasofree.bf",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "MotDePasseFort123!",
            "format": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Restaurant Faso Délices"
          },
          "address": {
            "type": "string",
            "example": "1200, Ouagadougou"
          },
          "phone": {
            "type": "string",
            "example": "+22670000001"
          },
          "latitude": {
            "type": "number",
            "example": 12.3714
          },
          "longitude": {
            "type": "number",
            "example": -1.5197
          },
          "brandId": {
            "type": "string",
            "description": "Marque (Brand) à laquelle rattacher cette agence"
          }
        },
        "required": [
          "name",
          "address",
          "phone",
          "latitude",
          "longitude"
        ]
      },
      "UpdateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nom du commerce"
          },
          "address": {
            "type": "string",
            "description": "Adresse"
          },
          "phone": {
            "type": "string",
            "description": "Téléphone"
          },
          "category": {
            "type": "string",
            "enum": [
              "RESTAURANT",
              "SUPERMARKET",
              "PHARMACY",
              "RETAIL",
              "BAKERY",
              "SERVICES"
            ],
            "description": "Catégorie du commerce"
          },
          "enableDelivery": {
            "type": "boolean",
            "description": "Activer la livraison"
          },
          "enablePickup": {
            "type": "boolean",
            "description": "Activer Click & Collect"
          },
          "enableDineIn": {
            "type": "boolean",
            "description": "Activer consommation sur place"
          },
          "hasOwnDrivers": {
            "type": "boolean",
            "description": "Utiliser ses propres livreurs"
          },
          "isOpen": {
            "type": "boolean",
            "description": "Statut d'ouverture"
          },
          "latitude": {
            "type": "number",
            "description": "Latitude"
          },
          "longitude": {
            "type": "number",
            "description": "Longitude"
          },
          "brandId": {
            "type": "string",
            "description": "Marque (Brand) de rattachement"
          }
        }
      },
...
    }
  }
}